Refactor: 패키지 구조 및 이름 리팩토링
