--- chat/src/main/java/org/bookwoori/chat/channelMessage/dto/request/ChannelMessageReactRequestDto.java
@@ -1,12 +1,0 @@
-package org.bookwoori.chat.channelMessage.dto.request;
-
-import org.bookwoori.chat.global.common.ActionType;
-import org.bookwoori.chat.global.common.EmojiType;
-
-public record ChannelMessageReactRequestDto(
-    String id,
-    EmojiType emoji,
-    ActionType action
-) {
-
-}
