--- chat/src/main/java/org/bookwoori/chat/channelMessage/dto/response/ChannelMessageListResponseDto.java
@@ -1,9 +1,0 @@
-package org.bookwoori.chat.channelMessage.dto.response;
-
-import java.util.List;
-
-public record ChannelMessageListResponseDto(
-    List<ChannelMessageItemDto> messages
-) {
-
-}
