--- chat/src/main/java/org/bookwoori/chat/directMessage/dto/response/DirectMessageListResponseDto.java
@@ -1,9 +1,0 @@
-package org.bookwoori.chat.directMessage.dto.response;
-
-import java.util.List;
-
-public record DirectMessageListResponseDto(
-    List<DirectMessageItemDto> messages
-) {
-
-}
