--- chat/src/main/java/org/bookwoori/chat/directMessage/dto/response/DirectMessageReactResponseDto.java
@@ -1,27 +1,0 @@
-package org.bookwoori.chat.directMessage.dto.response;
-
-import java.util.HashSet;
-import java.util.Set;
-import lombok.Builder;
-import org.bookwoori.chat.directMessage.domain.DirectMessage;
-import org.bookwoori.chat.global.common.EmojiType;
-
-@Builder
-public record DirectMessageReactResponseDto(
-    String id,
-    EmojiType emoji,
-    int emojiCount,
-    Set<Long> members
-) {
-
-    public static DirectMessageReactResponseDto from(DirectMessage directMessage) {
-        return DirectMessageReactResponseDto.builder()
-            .id(directMessage.getId())
-            .emoji(directMessage.getTargetEmoji())
-            .emojiCount(
-                directMessage.getReactions()
-                    .getOrDefault(directMessage.getTargetEmoji(), new HashSet<>()).size())
-            .members(directMessage.getReactions().get(directMessage.getTargetEmoji()))
-            .build();
-    }
-}
