+++ chat/src/main/java/org/bookwoori/chat/domain/channelMessage/dto/response/ChannelMessageItemDto.java
@@ -1,0 +1,25 @@
+package org.bookwoori.chat.domain.channelMessage.dto.response;
+
+import java.time.LocalDateTime;
+import lombok.Builder;
+import org.bookwoori.chat.domain.channelMessage.entity.ChannelMessage;
+
+@Builder
+public record ChannelMessageItemDto(
+    String id,
+    Long channelId,
+    Long memberId,
+    String content,
+    LocalDateTime createdAt
+) {
+
+    public static ChannelMessageItemDto from(ChannelMessage message) {
+        return ChannelMessageItemDto.builder()
+            .id(message.getId())
+            .channelId(message.getChannelId())
+            .memberId(message.getMemberId())
+            .content(message.getContent())
+            .createdAt(message.getCreatedAt())
+            .build();
+    }
+}
