+++ chat/src/main/java/org/bookwoori/chat/domain/channelMessage/dto/response/ChannelMessageListResponseDto.java
@@ -1,0 +1,9 @@
+package org.bookwoori.chat.domain.channelMessage.dto.response;
+
+import java.util.List;
+
+public record ChannelMessageListResponseDto(
+    List<ChannelMessageItemDto> messages
+) {
+
+}
