+++ chat/src/main/java/org/bookwoori/chat/domain/directMessage/dto/response/DirectMessageListResponseDto.java
@@ -1,0 +1,9 @@
+package org.bookwoori.chat.domain.directMessage.dto.response;
+
+import java.util.List;
+
+public record DirectMessageListResponseDto(
+    List<DirectMessageItemDto> messages
+) {
+
+}
