+++ chat/src/main/java/org/bookwoori/chat/global/kafka/MessageSender.java
@@ -2,16 +2,16 @@ package org.bookwoori.chat.global.kafka;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.bookwoori.chat.channelMessage.domain.ChannelMessage;
-import org.bookwoori.chat.channelMessage.dto.request.ChannelMessageReactRequestDto;
-import org.bookwoori.chat.channelMessage.dto.request.ChannelMessageReplyRequestDto;
-import org.bookwoori.chat.channelMessage.dto.request.ChannelMessageSendRequestDto;
-import org.bookwoori.chat.channelMessage.repository.ChannelMessageRepository;
-import org.bookwoori.chat.directMessage.domain.DirectMessage;
-import org.bookwoori.chat.directMessage.dto.request.DirectMessageReactRequestDto;
-import org.bookwoori.chat.directMessage.dto.request.DirectMessageReplyRequestDto;
-import org.bookwoori.chat.directMessage.dto.request.DirectMessageSendRequestDto;
-import org.bookwoori.chat.directMessage.repository.DirectMessageRepository;
+import org.bookwoori.chat.domain.channelMessage.entity.ChannelMessage;
+import org.bookwoori.chat.domain.channelMessage.dto.request.ChannelMessageReactRequestDto;
+import org.bookwoori.chat.domain.channelMessage.dto.request.ChannelMessageReplyRequestDto;
+import org.bookwoori.chat.domain.channelMessage.dto.request.ChannelMessageSendRequestDto;
+import org.bookwoori.chat.domain.channelMessage.repository.ChannelMessageRepository;
+import org.bookwoori.chat.domain.directMessage.entity.DirectMessage;
+import org.bookwoori.chat.domain.directMessage.dto.request.DirectMessageReactRequestDto;
+import org.bookwoori.chat.domain.directMessage.dto.request.DirectMessageReplyRequestDto;
+import org.bookwoori.chat.domain.directMessage.dto.request.DirectMessageSendRequestDto;
+import org.bookwoori.chat.domain.directMessage.repository.DirectMessageRepository;
 import org.bookwoori.chat.global.common.ActionType;
 import org.bookwoori.chat.global.common.EventType;
 import org.bookwoori.chat.global.exception.CustomException;
